Drafts handling added

+++ src/Database/Delete.hs
@@ -7,6 +7,7 @@ module Database.Delete where
 
 
 import qualified Data.ByteString as B
+import qualified Data.Text as T
 import qualified Database.PostgreSQL.Simple as PS
 
 import Action.Tags
@@ -17,18 +18,37 @@ import Action.Comments
 import Action.Draft
 import Execute.Types
 import Database.SqlValue
+import MonadTypes
+import qualified Database.PostgreSQL.Simple.Types as PSTy
+import qualified Exceptions as Ex
+import qualified Types as Ty
 
-class (PS.ToRow (Del s)) => DeleteSQL s where
+--import Execute.Permissions
+
+deleteThis' :: (MonadServer m, DeleteSQL s) => s -> Del s -> m [Int]
+deleteThis' s del = do
+    let (str, params) = deleteQuery s del
+    debugStr <- formatQuery str params
+    logDebug $ T.pack $ show debugStr
+
+    --withExceptionHandlers (Ex.defaultHandlers "deleteThis") $ do
+    ids <- fmap (map PSTy.fromOnly) $ query str params
+    logInfo $ "Deleted " <> dName s <> " with id = " <> showList ids
+    return ids
+
+
+
+class DeleteSQL s where
     type Del s :: *
     deleteQuery :: s -> Del s -> (PS.Query, [SqlValue])
-    dName :: s -> B.ByteString
+    dName :: s -> T.Text
 
 newtype DTag = DTag ()
 dummyDTag = DTag ()
 
 instance DeleteSQL DTag where
     type Del DTag = DeleteTag
-    deleteQuery _ dt = ("DELETE FROM news.tag WHERE tag_id = ?", [SqlValue $ _dt_tagId dt])
+    deleteQuery _ dt = ("DELETE FROM news.tag WHERE tag_id = ? RETURNING tag_id", [SqlValue $ _dt_tagId dt])
     dName _ = "tag"
 
 
@@ -37,7 +57,7 @@ dummyDCat = DCat ()
 
 instance DeleteSQL DCat where
     type Del DCat = DeleteCategory
-    deleteQuery _ dc = ("DELETE FROM news.category WHERE category_id = ?", [SqlValue $ _dc_catId dc])
+    deleteQuery _ dc = ("DELETE FROM news.category WHERE category_id = ? RETURNING category_id", [SqlValue $ _dc_catId dc])
     dName _ = "category"
 
 
@@ -46,7 +66,7 @@ dummyDAuthor = DAuthor ()
 
 instance DeleteSQL DAuthor where
     type Del DAuthor = DeleteAuthor
-    deleteQuery _ da = ("DELETE FROM news.author WHERE author_id = ?", [SqlValue $ _da_authorId da])
+    deleteQuery _ da = ("DELETE FROM news.author WHERE author_id = ? RETURNING author_id", [SqlValue $ _da_authorId da])
     dName _ = "author"
 
 newtype DUser = DUser ()
@@ -54,21 +74,31 @@ dummyDUser = DUser ()
 
 instance DeleteSQL DUser where
     type Del DUser = DeleteUser
-    deleteQuery _ du = ("DELETE FROM news.users WHERE user_id = ?", [SqlValue $ _du_userId du])
+    deleteQuery _ du = ("DELETE FROM news.users WHERE user_id = ? RETURNING user_id", [SqlValue $ _du_userId du])
     dName _ = "user"
 
 
 newtype DComment = DComment ()
 dummyDComment = DComment ()
 
+
+isAdmin :: Ty.User -> Bool
+--isAdmin u = Ty._u_admin u /= Just True
+isAdmin u = maybe False id $ Ty._u_admin u
+
+
+
 instance DeleteSQL DComment where
     type Del DComment = WithUser DeleteComment
-    deleteQuery _ (WithUser u dc) = (
-        "\
-\ DELETE FROM news.comment c \
-\ WHERE c.comment_id = ? AND EXISTS \
-\    (SELECT * FROM news.users u WHERE u.user_id = ? AND (u.is_admin = true OR u.user_id = c.user_id))"
-        , [SqlValue $ _dc_commentId dc, SqlValue u])
+    deleteQuery _ (WithUser u dc) = let
+        str = " DELETE FROM news.comment WHERE comment_id = ? "
+        userWhere = " AND user_id = ? "
+        returning = " RETURNING comment_id"
+        userParam = SqlValue $ Ty._u_id u
+        commentParam = SqlValue $ _dc_commentId dc
+        in  if   isAdmin u
+            then (str <> returning, [commentParam])
+            else (str <> userWhere <> returning, [commentParam, userParam])
 
 
     dName _ = "comment"
